requirements.txt
openai
requests
httpx
beautifulsoup4
lxml
pyyaml
python-dotenv
ddgs
